moved geo ip update to bootstrap script
