+++ deploy/Dockerfile-slim
@@ -1,5 +1,5 @@
-ARG NODEJS_IMAGE=node:18.16-alpine3.17
-ARG ALPINE_IMAGE=alpine:3.17
+ARG NODEJS_IMAGE=node:18.16.1-alpine3.18
+ARG ALPINE_IMAGE=alpine:3.18
 
 # # #
 # Cron build
