+++ src/_processor/index.ts
@@ -28,23 +28,8 @@ import { logger } from "./logger";
 import { startHealthz, updateLastNSQ } from "./healthz";
 import getPgPool from "../persistence/pg";
 import { notifyError, startErrorNotifier } from "../error-notifier";
-import { mmdbExists } from "../common/mmdb";
-
-const startGeoSync = async () => {
-  const pgPool = getPgPool();
-  const shouldSync = await pgPool.query("SELECT 1 FROM geoip LIMIT 1");
-  if (
-    !config.RETRACED_DISABLE_GEOSYNC &&
-    ((config.MAXMIND_GEOLITE2_LICENSE_KEY && shouldSync.rowCount === 0) ||
-      (config.GEO_USE_MMDB && !mmdbExists()))
-  ) {
-    updateGeoData();
-  }
-};
 
 startHealthz();
-
-startGeoSync();
 
 startErrorNotifier();
